fix: move proton go build to odpf/proton
